Generate stripe-node with breaking changes (#1491)
* Generate stripe-node with breaking changes

* Update CustomerSource enum

* Generate files

* Fix type in test

* Fix tests
