+++ lib/resources/CreditNotes.js
@@ -29,6 +29,12 @@ module.exports = StripeResource.extend({
     methodType: 'list',
   }),
 
+  listPreviewLineItems: stripeMethod({
+    method: 'GET',
+    path: '/preview/lines',
+    methodType: 'list',
+  }),
+
   preview: stripeMethod({
     method: 'GET',
     path: '/preview',
@@ -42,12 +48,6 @@ module.exports = StripeResource.extend({
   listLineItems: stripeMethod({
     method: 'GET',
     path: '/{creditNote}/lines',
-    methodType: 'list',
-  }),
-
-  listPreviewLineItems: stripeMethod({
-    method: 'GET',
-    path: '/preview/lines',
     methodType: 'list',
   }),
 });
