+++ lib/resources/Invoices.js
@@ -39,6 +39,12 @@ module.exports = StripeResource.extend({
     path: '/{invoice}/finalize',
   }),
 
+  listUpcomingLines: stripeMethod({
+    method: 'GET',
+    path: '/upcoming/lines',
+    methodType: 'list',
+  }),
+
   markUncollectible: stripeMethod({
     method: 'POST',
     path: '/{invoice}/mark_uncollectible',
@@ -73,12 +79,6 @@ module.exports = StripeResource.extend({
   listLineItems: stripeMethod({
     method: 'GET',
     path: '/{invoice}/lines',
-    methodType: 'list',
-  }),
-
-  listUpcomingLineItems: stripeMethod({
-    method: 'GET',
-    path: '/upcoming/lines',
     methodType: 'list',
   }),
 });
