+++ lib/resources/Issuing/Cards.js
@@ -28,9 +28,4 @@ module.exports = StripeResource.extend({
     path: '',
     methodType: 'list',
   }),
-
-  retrieveDetails: stripeMethod({
-    method: 'GET',
-    path: '/{card}/details',
-  }),
 });
