+++ types/2020-08-27/CreditNoteLineItems.d.ts
@@ -126,116 +126,5 @@ declare module 'stripe' {
        */
       expand?: Array<string>;
     }
-
-    interface CreditNoteLineItemListPreviewParams extends PaginationParams {
-      /**
-       * ID of the invoice.
-       */
-      invoice: string;
-
-      /**
-       * The integer amount in cents (or local equivalent) representing the total amount of the credit note.
-       */
-      amount?: number;
-
-      /**
-       * The integer amount in cents (or local equivalent) representing the amount to credit the customer's balance, which will be automatically applied to their next invoice.
-       */
-      credit_amount?: number;
-
-      /**
-       * Specifies which fields in the response should be expanded.
-       */
-      expand?: Array<string>;
-
-      /**
-       * Line items that make up the credit note.
-       */
-      lines?: Array<CreditNoteLineItemListPreviewParams.Line>;
-
-      /**
-       * The credit note's memo appears on the credit note PDF.
-       */
-      memo?: string;
-
-      /**
-       * Set of [key-value pairs](https://stripe.com/docs/api/metadata) that you can attach to an object. This can be useful for storing additional information about the object in a structured format. Individual keys can be unset by posting an empty value to them. All keys can be unset by posting an empty value to `metadata`.
-       */
-      metadata?: Stripe.MetadataParam;
-
-      /**
-       * The integer amount in cents (or local equivalent) representing the amount that is credited outside of Stripe.
-       */
-      out_of_band_amount?: number;
-
-      /**
-       * Reason for issuing this credit note, one of `duplicate`, `fraudulent`, `order_change`, or `product_unsatisfactory`
-       */
-      reason?: CreditNoteLineItemListPreviewParams.Reason;
-
-      /**
-       * ID of an existing refund to link this credit note to.
-       */
-      refund?: string;
-
-      /**
-       * The integer amount in cents (or local equivalent) representing the amount to refund. If set, a refund will be created for the charge associated with the invoice.
-       */
-      refund_amount?: number;
-    }
-
-    namespace CreditNoteLineItemListPreviewParams {
-      interface Line {
-        /**
-         * The line item amount to credit. Only valid when `type` is `invoice_line_item`.
-         */
-        amount?: number;
-
-        /**
-         * The description of the credit note line item. Only valid when the `type` is `custom_line_item`.
-         */
-        description?: string;
-
-        /**
-         * The invoice line item to credit. Only valid when the `type` is `invoice_line_item`.
-         */
-        invoice_line_item?: string;
-
-        /**
-         * The line item quantity to credit.
-         */
-        quantity?: number;
-
-        /**
-         * The tax rates which apply to the credit note line item. Only valid when the `type` is `custom_line_item`.
-         */
-        tax_rates?: Stripe.Emptyable<Array<string>>;
-
-        /**
-         * Type of the credit note line item, one of `invoice_line_item` or `custom_line_item`
-         */
-        type: Line.Type;
-
-        /**
-         * The integer unit amount in cents (or local equivalent) of the credit note line item. This `unit_amount` will be multiplied by the quantity to get the full amount to credit for this line item. Only valid when `type` is `custom_line_item`.
-         */
-        unit_amount?: number;
-
-        /**
-         * Same as `unit_amount`, but accepts a decimal value in cents (or local equivalent) with at most 12 decimal places. Only one of `unit_amount` and `unit_amount_decimal` can be set.
-         */
-        unit_amount_decimal?: string;
-      }
-
-      namespace Line {
-        type Type = 'custom_line_item' | 'invoice_line_item';
-      }
-
-      type Reason =
-        | 'duplicate'
-        | 'fraudulent'
-        | 'order_change'
-        | 'product_unsatisfactory';
-    }
   }
 }
