+++ types/2020-08-27/CreditNotes.d.ts
@@ -338,6 +338,117 @@ declare module 'stripe' {
       invoice?: string;
     }
 
+    interface CreditNoteListPreviewLineItemsParams extends PaginationParams {
+      /**
+       * ID of the invoice.
+       */
+      invoice: string;
+
+      /**
+       * The integer amount in cents (or local equivalent) representing the total amount of the credit note.
+       */
+      amount?: number;
+
+      /**
+       * The integer amount in cents (or local equivalent) representing the amount to credit the customer's balance, which will be automatically applied to their next invoice.
+       */
+      credit_amount?: number;
+
+      /**
+       * Specifies which fields in the response should be expanded.
+       */
+      expand?: Array<string>;
+
+      /**
+       * Line items that make up the credit note.
+       */
+      lines?: Array<CreditNoteListPreviewLineItemsParams.Line>;
+
+      /**
+       * The credit note's memo appears on the credit note PDF.
+       */
+      memo?: string;
+
+      /**
+       * Set of [key-value pairs](https://stripe.com/docs/api/metadata) that you can attach to an object. This can be useful for storing additional information about the object in a structured format. Individual keys can be unset by posting an empty value to them. All keys can be unset by posting an empty value to `metadata`.
+       */
+      metadata?: Stripe.MetadataParam;
+
+      /**
+       * The integer amount in cents (or local equivalent) representing the amount that is credited outside of Stripe.
+       */
+      out_of_band_amount?: number;
+
+      /**
+       * Reason for issuing this credit note, one of `duplicate`, `fraudulent`, `order_change`, or `product_unsatisfactory`
+       */
+      reason?: CreditNoteListPreviewLineItemsParams.Reason;
+
+      /**
+       * ID of an existing refund to link this credit note to.
+       */
+      refund?: string;
+
+      /**
+       * The integer amount in cents (or local equivalent) representing the amount to refund. If set, a refund will be created for the charge associated with the invoice.
+       */
+      refund_amount?: number;
+    }
+
+    namespace CreditNoteListPreviewLineItemsParams {
+      interface Line {
+        /**
+         * The line item amount to credit. Only valid when `type` is `invoice_line_item`.
+         */
+        amount?: number;
+
+        /**
+         * The description of the credit note line item. Only valid when the `type` is `custom_line_item`.
+         */
+        description?: string;
+
+        /**
+         * The invoice line item to credit. Only valid when the `type` is `invoice_line_item`.
+         */
+        invoice_line_item?: string;
+
+        /**
+         * The line item quantity to credit.
+         */
+        quantity?: number;
+
+        /**
+         * The tax rates which apply to the credit note line item. Only valid when the `type` is `custom_line_item`.
+         */
+        tax_rates?: Stripe.Emptyable<Array<string>>;
+
+        /**
+         * Type of the credit note line item, one of `invoice_line_item` or `custom_line_item`
+         */
+        type: Line.Type;
+
+        /**
+         * The integer unit amount in cents (or local equivalent) of the credit note line item. This `unit_amount` will be multiplied by the quantity to get the full amount to credit for this line item. Only valid when `type` is `custom_line_item`.
+         */
+        unit_amount?: number;
+
+        /**
+         * Same as `unit_amount`, but accepts a decimal value in cents (or local equivalent) with at most 12 decimal places. Only one of `unit_amount` and `unit_amount_decimal` can be set.
+         */
+        unit_amount_decimal?: string;
+      }
+
+      namespace Line {
+        type Type = 'custom_line_item' | 'invoice_line_item';
+      }
+
+      type Reason =
+        | 'duplicate'
+        | 'fraudulent'
+        | 'order_change'
+        | 'product_unsatisfactory';
+    }
+
     interface CreditNotePreviewParams {
       /**
        * ID of the invoice.
@@ -510,6 +621,14 @@ declare module 'stripe' {
       list(options?: RequestOptions): ApiListPromise<Stripe.CreditNote>;
 
       /**
+       * When retrieving a credit note preview, you'll get a lines property containing the first handful of those items. This URL you can retrieve the full (paginated) list of line items.
+       */
+      listPreviewLineItems(
+        params: CreditNoteListPreviewLineItemsParams,
+        options?: RequestOptions
+      ): ApiListPromise<Stripe.CreditNoteLineItem>;
+
+      /**
        * Get a preview of a credit note without creating it.
        */
       preview(
@@ -540,14 +659,6 @@ declare module 'stripe' {
       ): ApiListPromise<Stripe.CreditNoteLineItem>;
       listLineItems(
         id: string,
-        options?: RequestOptions
-      ): ApiListPromise<Stripe.CreditNoteLineItem>;
-
-      /**
-       * When retrieving a credit note preview, you'll get a lines property containing the first handful of those items. This URL you can retrieve the full (paginated) list of line items.
-       */
-      listPreviewLineItems(
-        params: CreditNoteLineItemListPreviewParams,
         options?: RequestOptions
       ): ApiListPromise<Stripe.CreditNoteLineItem>;
     }
