+++ types/2020-08-27/CustomerSources.d.ts
@@ -17,6 +17,8 @@ declare module 'stripe' {
        * Set of [key-value pairs](https://stripe.com/docs/api/metadata) that you can attach to an object. This can be useful for storing additional information about the object in a structured format. Individual keys can be unset by posting an empty value to them. All keys can be unset by posting an empty value to `metadata`.
        */
       metadata?: Stripe.MetadataParam;
+
+      validate?: boolean;
     }
 
     interface CustomerSourceRetrieveParams {
