+++ types/2020-08-27/WebhookEndpoints.d.ts
@@ -328,7 +328,6 @@ declare module 'stripe' {
         | 'issuing_transaction.updated'
         | 'mandate.updated'
         | 'order.created'
-        | 'order.payment_failed'
         | 'payment_intent.amount_capturable_updated'
         | 'payment_intent.canceled'
         | 'payment_intent.created'
@@ -414,8 +413,6 @@ declare module 'stripe' {
         | 'topup.reversed'
         | 'topup.succeeded'
         | 'transfer.created'
-        | 'transfer.failed'
-        | 'transfer.paid'
         | 'transfer.reversed'
         | 'transfer.updated'
         | 'treasury.credit_reversal.created'
@@ -444,7 +441,6 @@ declare module 'stripe' {
         | 'treasury.outbound_transfer.returned'
         | 'treasury.received_credit.created'
         | 'treasury.received_credit.failed'
-        | 'treasury.received_credit.reversed'
         | 'treasury.received_credit.succeeded'
         | 'treasury.received_debit.created';
     }
@@ -592,7 +588,6 @@ declare module 'stripe' {
         | 'issuing_transaction.updated'
         | 'mandate.updated'
         | 'order.created'
-        | 'order.payment_failed'
         | 'payment_intent.amount_capturable_updated'
         | 'payment_intent.canceled'
         | 'payment_intent.created'
@@ -678,8 +673,6 @@ declare module 'stripe' {
         | 'topup.reversed'
         | 'topup.succeeded'
         | 'transfer.created'
-        | 'transfer.failed'
-        | 'transfer.paid'
         | 'transfer.reversed'
         | 'transfer.updated'
         | 'treasury.credit_reversal.created'
@@ -708,7 +701,6 @@ declare module 'stripe' {
         | 'treasury.outbound_transfer.returned'
         | 'treasury.received_credit.created'
         | 'treasury.received_credit.failed'
-        | 'treasury.received_credit.reversed'
         | 'treasury.received_credit.succeeded'
         | 'treasury.received_debit.created';
     }
