+++ types/test/typescriptTest.ts
@@ -53,7 +53,7 @@ stripe.setHost('host', 'port', 'protocol');
   };
   const customer: Stripe.Customer = await stripe.customers.create(params, opts);
 
-  const address: Stripe.Address | null = customer.address;
+  const address: Stripe.Address | null | undefined = customer.address;
 
   if (!address) return;
   const city: string | null = address.city;
